Update RMS measurement max voltage, QoL changes
This change contains a new max voltage of 2000V to work around the current transformer setup. It also uses the LED to show when logging, and to always sync the RTC. The code will, for now, log the RMS and DC bias values to the SD card immediately when powered and will continue doing so until it is powered off (will be changed later). The datalogging code is commented out to prevent file opening conflict, but that can be handled later as well.